Skip test that unmounts/remounts metadata partition
Aconfig flagging is migrating to a new storage backend that makes heavy
use of the /metadata partition. It is causing failures in tests like
vts_libsnapshot_test, which unmount and remount this partition. The
failed unmounts are due to the partition being busy.

The test should be reworked to not unmount and remount the partition;
linked bug in GSKIP_TEST for followup.

Bug: 349807180
Bug: 350715463
Test: m vts_libsnapshot_test
Change-Id: Ie4f8cadcad189b13b9eba1eb15c5251002e8138a

Former-commit-id: ef0f5bbcf117f7abf5b2c671568fb23b7200f41f