fs_mgr: check available space before creating zram backing device
This CL adds a check to ensure that there is enough free space on
the device before creating the zram backing device.
If there is not enough space, the zram backing device will not be
created.

Adding a system property 'ro.zram_backing_device_min_free_mb' to
define the minimum free space reserved for booting.
If the property is not set, skip the available space check.

Bug: 322821637
Bug: 326877642
Test: Fill up userdata, enlarge zram backing device size on next
boot.

Change-Id: I135b4beee2a8a21f8214f6579368a8d0d8792255

Former-commit-id: 6c277195abb1118a135eed284dff83f4ba3b70a6